Pull it baby!

+++ modules/Your_Account/includes/cookiecheck.php
@@ -159,14 +159,23 @@ function ShowCookiesRedirect()
 /*************************************************************************************/
 function ShowCookies() 
 {
-  global $ya_config,$module_name;
+  global $ya_config,$module_name,$prefix,$user,$username,$CookieArray,$cookie;
   include_once(NUKE_BASE_DIR.'header.php');
-  //Show_CNBYA_menu(); # removed excessive menu system from show cookies
+
   OpenTable();
 
-    //$CookieArray = explode('; ', $_SERVER['HTTP_COOKIE']);
     $CookieArray = $_COOKIE;
 
+	if(isset($cookie[0]))
+    $r_uid = $cookie[0];
+	else
+	$r_uid = '1';
+	
+	if(isset($cookie[1]))
+    $r_username = $cookie[1];
+	else
+    $r_username = 'Anonymous';
+	
     echo "<table width=\"100%\" cellspacing=\"0\" cellpadding=\"5\" border=\"0\"><tr>";
     echo "<form action=\"modules.php?name=$module_name&amp;op=DeleteCookies\" method=\"post\">";
     echo "<td style=\"padding-bottom: 0px;\" colspan=\"2\">";
@@ -175,7 +184,7 @@ function ShowCookies()
 
 	echo '<fieldset style="border-color: gold; border-width: '.$fieldset_border_width.'; border-style: solid;">';
     echo '<legend align="left" id="Legend5" runat="server" visible="true" style="width:auto; margin: 0px; font-weight: bold;"><img 
-	src="modules/'.$module_name.'/images/warning.png" align="left" width="20" height="20"></strong>&nbsp;<span style="color: red">Delete Cookies</span>&nbsp;</legend>';
+	src="modules/'.$module_name.'/images/warning.png" align="left" width="20" height="20"></strong>&nbsp;<span style="color: red">'.$r_username.'\'s Cookies</span>&nbsp;</legend>';
 	
 	echo "<span class=\"content\">"._YA_DELCOOKIEINFO1."</span></td></tr><tr><td width=\"100%\"></fieldset>";
 
@@ -185,16 +194,20 @@ function ShowCookies()
     
 	if (is_array($CookieArray) && !empty($CookieArray)) 
 	{
-        //while(list($cName,$cValue) = each($CookieArray)) # PHP 8.1 Fix
 		foreach ($CookieArray as $cName => $cValue) 
 		{
-            $cName     = str_replace(" ","",$cName); 
-            if (empty($cValue)) $cValue = "(empty)";
-            $cMore = substr("$cValue", 0, 37)." <span style=\"color: white;\">...✂</span>";
-			//$cMore     = substr("$cValue", 36, 1);
+            $cName = str_replace(" ","",$cName); 
+            
+			if (empty($cValue)) $cValue = "(empty)";
+            
+			$cMore = substr("$cValue", 0, 37)." <span style=\"color: white;\">...✂</span>";
+
             if (!empty($cMore)) 
             $cValue = substr("$cValue", 0, 37)." <span style=\"color: white;\">...✂</span>";
             
+			if($cName === 'PHPSESSID')
+			echo "<tr><td align=\"left\" nowrap=\"nowrap\"><span style=\"color:crimson;\">$cName</span>&nbsp;&nbsp;&nbsp;</td><td width=\"100%\" align=\"left\"><span style=\"color:yellow;\">🍪 $cValue</span></td></tr>";
+			else
 			echo "<tr><td align=\"left\" nowrap=\"nowrap\"><span style=\"color:lime;\">$cName</span>&nbsp;&nbsp;&nbsp;</td><td width=\"100%\" align=\"left\"><span style=\"color:yellow;\">🍪 $cValue</span></td></tr>";
         }
             
@@ -216,22 +229,26 @@ function ShowCookies()
 function DeleteCookies() {
 global $ya_config,$module_name,$prefix,$user,$username,$CookieArray,$cookie;
 include_once(NUKE_BASE_DIR.'header.php');
-//Show_CNBYA_menu(); # Remove excessive menu from delete cookies
 OpenTable();
-
-    $r_uid        = $cookie[0];
-    $r_username    = $cookie[1];
-    echo $r_username;
-    echo $r_uid;
-    echo $username;
+    
+	if(isset($cookie[0]))
+    $r_uid = $cookie[0];
+	else
+	$r_uid = '1';
+	
+	if(isset($cookie[1]))
+    $r_username = $cookie[1];
+	else
+    $r_username = 'Anonymous';
 
     $CookieArray = $_COOKIE;
-    $db->sql_query("DELETE FROM ".$prefix."_session WHERE uname='$r_username'");
-    $db->sql_query("OPTIMIZE TABLE ".$prefix."_session");
-//    $db->sql_query("DELETE FROM     ".$prefix."_bbsessions WHERE session_user_id='$r_uid'");
-//    $db->sql_query("OPTIMIZE TABLE ".$prefix."_bbsessions");
-
-    echo "<table width=\"100%\" cellspacing=\"0\" cellpadding=\"5\" border=\"0\"><tr>";
+	
+	if(isset($r_username)):
+      $db->sql_query("DELETE FROM ".$prefix."_session WHERE uname='$r_username'");
+      $db->sql_query("OPTIMIZE TABLE ".$prefix."_session");
+    endif;
+    
+	echo "<table width=\"100%\" cellspacing=\"0\" cellpadding=\"5\" border=\"0\"><tr>";
     echo "<form action=\"modules.php?name=$module_name&amp;op=ShowCookies\" method=\"post\">";
     echo "<td colspan=\"2\"><img src=\"modules/$module_name/images/warning.png\" align=\"left\" width=\"40\" height=\"40\">";
   
@@ -240,8 +257,9 @@ OpenTable();
     echo "<table cellspacing=\"0\" cellpadding=\"5\" border=\"1\" align=\"left\"><tr><td colspan=\"2\">";
     echo "<span class=\"title\">"._YA_CURRENTCOOKIE."</span></td></tr>";
     echo "<tr><td nowrap=\"nowrap\"><strong>"._YA_COOKIENAME."</strong></td><td width=\"100%\"><strong>"._YA_COOKIESTAT."</strong></td></tr>";
+
     if (is_array($CookieArray) && !empty($CookieArray)) {
-        //while(list($cName,$cValue) = each($CookieArray)) { # PHP 8.1 Fix
+
 		foreach ($CookieArray as $cName => $cValue)
 		{	
             $cName = str_replace(" ","",$cName);
@@ -249,6 +267,7 @@ OpenTable();
             setcookie("$cName","1",time()-604800,"");                          // Directory only path
             setcookie("$cName","2",time()-604800,"/");                         // Site wide path
             setcookie("$cName","3",time()-604800,"$ya_config[cookiepath]");    // Configured path
+
             echo "<tr><td align=\"left\" nowrap=\"nowrap\">$cName</td><td width=\"100%\" align=\"left\">"._YA_COOKIEDEL2."</td></tr>";
             unset($cName);
         }
@@ -259,7 +278,6 @@ OpenTable();
     echo "</td><td valign=\"top\"><input type=\"submit\" name=\"submit\" value='"._YA_COOKIESHOWALL."'></td></form></tr></table>";
     }
 
-# menelaos: these lines need some more study: which are usefull, which are not
 unset($user);
 unset($cookie);
 
@@ -273,7 +291,7 @@ if(isset($_SESSION)):
   session_destroy();
 endif; 
 
-if( isset($_COOKIE[session_name()])):
+if(isset($_COOKIE[session_name()])):
   unset($_COOKIE[session_name()]);
 endif;
 # menelaos: these lines need some more study: which are usefull, which are not
@@ -281,5 +299,3 @@ endif;
 CloseTable();
 include_once(NUKE_BASE_DIR.'footer.php');
 }
-
-?>
